requirements.txt
click==8.0.4
Flask==2.0.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
Werkzeug==2.0.3
eth-brownie==1.18.1
python-dotenv==0.16.0